[REF] Remove duplicate functions (#502)
