+++ xcp_d/interfaces/report_core.py
@@ -9,6 +9,8 @@ from pathlib import Path
 
 from niworkflows.reports.core import Report as _Report
 
+from xcp_d.utils.concantenation import _getsesid
+
 
 class Report(_Report):
     """A modified form of niworkflows' core Report object."""
@@ -163,18 +165,3 @@ def generate_reports(subject_list,
 
         print('Reports generated successfully')
     return errno
-
-
-def _getsesid(filename):
-    """Get session ID from filename."""
-    import os
-    ses_id = None
-    filex = os.path.basename(filename)
-
-    file_id = filex.split('_')
-    for k in file_id:
-        if 'ses' in k:
-            ses_id = k.split('-')[1]
-            break
-
-    return ses_id
