+++ xcp_d/utils/concantenation.py
@@ -16,6 +16,7 @@ from nipype.interfaces.ants import ApplyTransforms
 from templateflow.api import get as get_template
 
 from xcp_d.utils.plot import plot_svgx
+from xcp_d.utils.qcmetrics import compute_dvars
 from xcp_d.utils.utils import get_transformfile
 from xcp_d.utils.write_save import read_ndata
 
@@ -311,27 +312,6 @@ def concatenate_nifti(subid, fmridir, outputdir, ses=None, work_dir=None):
             shutil.copy(bb1ref, bboldref)
 
 
-def compute_dvars(datat):
-    """Compute standard DVARS.
-
-    Parameters
-    ----------
-    datat : numpy.ndarray
-        The data matrix fromw hich to calculate DVARS.
-        Ordered as vertices by timepoints.
-
-    Returns
-    -------
-    numpy.ndarray
-        The calculated DVARS array.
-        A (timepoints,) array.
-    """
-    firstcolumn = np.zeros((datat.shape[0]))[..., None]
-    datax = np.hstack((firstcolumn, np.diff(datat)))
-    datax_ss = np.sum(np.square(datax), axis=0) / datat.shape[0]
-    return np.sqrt(datax_ss)
-
-
 def concatenate_cifti(subid, fmridir, outputdir, ses=None, work_dir=None):
     """Concatenate CIFTI files along the time dimension.
 
@@ -549,6 +529,8 @@ def get_segfile(bold_file):
 def _t12native(fname):
     """Select T1w-to-scanner transform associated with a given BOLD file.
 
+    TODO: Update names and refactor
+
     Parameters
     ----------
     fname : str
@@ -600,8 +582,6 @@ def get_cifti_tr(cifti_file):
     float
         The TR of the CIFTI file.
     """
-    import nibabel as nb
-
     ciaxis = nb.load(cifti_file).header.get_axis(0)
     return ciaxis.step
